Move todolist hooks to each page hooks folder
