+++ client/app/utils/hooks/index.ts
@@ -2,5 +2,3 @@
 
 export * from './drag/useDragFirst'
 export * from './drag/useDragDndKit'
-export * from './todolist/useTodolist'
-export * from './todolist/useTodolistEditModal'
